requirements.txt
absl-py==0.8.1
appnope==0.1.0
astor==0.8.0
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bokeh==1.3.4
bz2file==0.98
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cycler==0.10.0
decorator==4.4.1
dill==0.3.1.1
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.0.2
django-filter==2.2.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djoser==2.0.3
dopamine-rl==2.0.5
Flask==1.1.1
future==0.18.1
gast==0.2.2
gevent==1.4.0
gin-config==0.2.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-pasta==0.1.7
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.24.3
gunicorn==19.9.0
gym==0.15.3
h5py==2.10.0
httplib2==0.14.0
idna==2.8
intervaltree==3.0.2
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.2.0
kiwisolver==1.1.0
librosa==0.7.1
llvmlite==0.30.0
magenta==1.1.7
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mesh-tensorflow==0.1.4
mido==1.2.6
mir-eval==0.5
mpmath==1.1.0
networkx==1.8.1
numba==0.46.0
numpy==1.17.3
oauth2client==4.1.3
opencv-python==4.1.1.26
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pretty-midi==0.2.8
promise==2.2.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pyglet==1.3.2
Pygments==2.4.2
pygtrie==2.3.2
PyJWT==1.7.1
pyparsing==2.4.2
pypng==0.0.20
python-dateutil==2.8.0
python-rtmidi==1.1.2
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sk-video==1.1.10
sonnet==0.1.6
sortedcontainers==2.1.0
SoundFile==0.10.2
sox==1.3.7
sqlparse==0.3.0
sympy==1.4
tensor2tensor==1.14.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-datasets==1.3.0
tensorflow-estimator==1.15.1
tensorflow-gan==2.0.0
tensorflow-hub==0.6.0
tensorflow-metadata==0.15.0
tensorflow-probability==0.7.0rc0
termcolor==1.1.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
whitenoise==4.1.4
wrapt==1.11.2
